argos-backend/requirements.txt
langchain
crewai
scrapy
ethers.py
web3
requests
beautifulsoup4
textblob
python-dotenv
openai
